モデルバリデーション修正とTodaymealモデルの追加（Progress rate: 100％、Branch name: feature/fix_model_validates）

+++ src/db/migrate/20210912133104_create_tranings.rb
@@ -5,8 +5,8 @@ class CreateTranings < ActiveRecord::Migration[5.2]
       t.string :traning_name
       t.string :sub_bodypart
       t.string :bodypart
-      t.integer :traning_weight
-      t.integer :traning_reps
+      t.float :traning_weight
+      t.float :traning_reps
       t.string :traning_note
 
       t.timestamps
